feat: implement login and change password pages with validation and styling
- Removed old login component and created a new one with improved structure and styles.
- Added change password component with validation for current and new passwords.
- Introduced CSS files for both login and change password pages to enhance UI.
- Implemented currency formatting for service registration.
- Created a service registration form with validation for required fields.
