+++ src/App.js
@@ -2,11 +2,11 @@ import "./App.css";
 
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import Dashboard from "./home/pages/Dashboard.jsx";
-import Login from "./auth/pages/login";
-import AlterarSenha from "./auth/pages/alterarsenha";
-import CadastroServico from "./auth/pages/CadastroServico";
-import SectionPage from "./home/pages/SectionPage";
-import ProtectedRoute from "./home/components/ProtectedRoute";
+import Login from "./auth/pages/login/login.jsx";
+import ChangePassword from "./auth/pages/changepassword/ChangePassword.jsx";
+import RegisterService from "./form/pages/registerservice/RegisterService.jsx";
+import SectionPage from "./home/pages/SectionPage.jsx";
+import ProtectedRoute from "./home/components/ProtectedRoute.jsx";
 import CadastroUsuario from "./auth/pages/CadastroUsuario.jsx";
 
 function App() {
@@ -19,9 +19,18 @@ function App() {
             <Route path="/dashboard" element={<Dashboard />} />
             <Route path="/login" element={<Login />} />
             <Route path="/cadastroUsuario" element={<CadastroUsuario />} />
-            <Route path="/alterarSenha" element={<AlterarSenha />} />
-            <Route path="/cadastroServico" element={<CadastroServico />} />
+            <Route path="/cadastro-usuario" element={<CadastroUsuario />} />
+            <Route path="/alterarSenha" element={<ChangePassword />} />
+            <Route path="/alterar-senha" element={<ChangePassword />} />
+            <Route path="/cadastroServico" element={<RegisterService />} />
+            <Route path="/cadastro-servico" element={<RegisterService />} />
+            <Route path="/clientes/novo" element={<CadastroUsuario />} />
+            <Route
+              path="/servicos-estoque/novo-servico"
+              element={<RegisterService />}
+            />
 
+            {/* Fallbacks temporarios para areas do dashboard que ainda nao possuem telas dedicadas. */}
             <Route
               path="/perfil"
               element={
@@ -30,7 +39,7 @@ function App() {
                   title="Perfil e preferências de acesso"
                   description="Este é um exemplo de tela para a futura integração com perfil autenticado, preferências e troca de credenciais."
                   ctaLabel="Voltar ao painel"
-                  ctaPath="/"
+                  ctaPath="/dashboard"
                 />
               }
             />
@@ -42,7 +51,7 @@ function App() {
                   title="Relatórios"
                   description="Visualize indicadores de desempenho, produtividade da equipe e evolução das ordens de serviço."
                   ctaLabel="Voltar ao painel"
-                  ctaPath="/"
+                  ctaPath="/dashboard"
                 />
               }
             />
@@ -54,7 +63,7 @@ function App() {
                   title="Configurações"
                   description="Ajuste parâmetros da operação, preferências de uso e regras internas do FaseCerta."
                   ctaLabel="Voltar ao painel"
-                  ctaPath="/"
+                  ctaPath="/dashboard"
                 />
               }
             />
@@ -66,7 +75,7 @@ function App() {
                   title="Gestão de Clientes"
                   description="Área demonstrativa para a futura navegação do módulo de clientes."
                   ctaLabel="Voltar ao painel"
-                  ctaPath="/"
+                  ctaPath="/dashboard"
                 />
               }
             />
@@ -78,7 +87,19 @@ function App() {
                   title="Serviços e Estoque"
                   description="Área demonstrativa para a futura navegação do módulo de serviços e estoque."
                   ctaLabel="Voltar ao painel"
-                  ctaPath="/"
+                  ctaPath="/dashboard"
+                />
+              }
+            />
+            <Route
+              path="/servicos-estoque/novo-item"
+              element={
+                <SectionPage
+                  eyebrow="Fallback temporário"
+                  title="Novo Item"
+                  description="Ainda não existe uma tela dedicada para cadastro de itens no estoque. Esta rota permanece como placeholder até a implementação do formulário real."
+                  ctaLabel="Voltar ao painel"
+                  ctaPath="/dashboard"
                 />
               }
             />
@@ -90,7 +111,19 @@ function App() {
                   title="Pedidos / Ordens de Serviço"
                   description="Área demonstrativa para a futura navegação do módulo de pedidos e ordens de serviço."
                   ctaLabel="Voltar ao painel"
-                  ctaPath="/"
+                  ctaPath="/dashboard"
+                />
+              }
+            />
+            <Route
+              path="/os/novo"
+              element={
+                <SectionPage
+                  eyebrow="Fallback temporário"
+                  title="Criar OS"
+                  description="Ainda não existe uma página de formulário dedicada para abertura de ordem de serviço. Esta rota foi mantida como placeholder até a tela real ser criada."
+                  ctaLabel="Voltar ao painel"
+                  ctaPath="/dashboard"
                 />
               }
             />
@@ -103,9 +136,21 @@ function App() {
                     title="Financeiro"
                     description="Acesso permitido somente para o perfil administrativo / gestor. A ocultação visual não substitui a validação real no backend."
                     ctaLabel="Voltar ao painel"
-                    ctaPath="/"
+                    ctaPath="/dashboard"
                   />
                 </ProtectedRoute>
+              }
+            />
+            <Route
+              path="/financeiro/despesas/nova"
+              element={
+                <SectionPage
+                  eyebrow="Fallback temporário"
+                  title="Cadastrar Despesa"
+                  description="Ainda não existe uma tela financeira específica para lançamento de despesas. Esta rota foi conectada como placeholder funcional."
+                  ctaLabel="Voltar ao painel"
+                  ctaPath="/dashboard"
+                />
               }
             />
             <Route path="*" element={<Navigate to="/" replace />} />
