+++ src/auth/pages/changepassword/ChangePassword.jsx
@@ -1,9 +1,12 @@
 import React, { useState, useEffect } from "react";
 // import { FaEye, FaEyeSlash, FaMoon } from "react-icons/fa";
-import "./alterarsenha.css";
+import { Link, useNavigate } from "react-router-dom";
 import { FaArrowLeft } from "react-icons/fa";
+import "./ChangePassword.css";
 
 const AlterarSenha = () => {
+  const navigate = useNavigate();
+
   // Estados para os valores dos inputs
   const [formData, setFormData] = useState({
     currentPassword: "",
@@ -105,7 +108,11 @@ const AlterarSenha = () => {
       <main className="change-password-container">
         <section className="change-password-card">
           <div className="card-header">
-            <button className="back-button">
+            <button
+              type="button"
+              className="back-button"
+              onClick={() => navigate("/dashboard")}
+            >
               <FaArrowLeft size={20} />
             </button>
             <h1>Alterar Senha</h1>
@@ -223,9 +230,9 @@ const AlterarSenha = () => {
           </form>
 
           <div className="card-footer">
-            <a href="#" className="recovery-link">
+            <Link to="/login" className="recovery-link">
               Esqueceu sua senha? <strong>Recuperar senha</strong>
-            </a>
+            </Link>
           </div>
         </section>
       </main>
